pyproject.toml
[project]
name = "lagchain-tutorial"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flax>=0.12.3",
    "grandalf>=0.8",
    "huggingface-hub>=0.36.2",
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.11",
    "langchain-huggingface>=1.2.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "scikit-learn>=1.8.0",
    "sentence-transformers>=5.2.2",
    "streamlit>=1.54.0",
    "tensorflow>=2.20.0",
    "textblob>=0.19.0",
    "tf-keras>=2.20.1",
    "torch>=2.10.0",
    "transformers>=4.57.6",
]
